Move all waiting functions to wait.py
