+++ integration-testing/test/test_heterogenous_validators.py
@@ -3,8 +3,9 @@ import conftest
 import contextlib
 
 from rnode_testing.rnode import start_bootstrap, create_peer
-from rnode_testing.wait import wait_for, node_started
-from rnode_testing.network import (
+from rnode_testing.wait import (
+    wait_for,
+    node_started,
     wait_for_approved_block_received_handler_state,
 )
